refactor vpn config creation
